:memo: Add docs for parameter dependencies
